초기 데이터 설계
