+++ Data_design/CSV/user_list.csv
@@ -1,3 +1,4 @@
 ﻿user_name,user_birth,user_sex,user_num,user_email,user_del,user_photo,user_rent
 김길동,19980101,0,010-1234-5678,rlfehd@naver.com,0,,0
 홍길동,19990101,0,010-4444-5555,ghd@naver.com,0,,0
+박길동,19990101,0,010-1234-5678,ghd@naver.com,0,,0
